refactoring of ObjectJc: now only one definition, also in ObjectSimple with all elements possible Tested all.
* Base/Exception.h: only some defines to clarify situation.
* Base/Handle_ptr64_TimeMeas_emC.c adapt
* Base/ObjectRefl_emC.c ctor only in ObjectSimple_emC.c
* Base/ObjectRefl_emC.h: cleaned, some definitions only in ObjectSimle_emC.h
* Base/ObjectSimple_emC.*: enhanced with all capablilities of ObjectJc
* Base/applstdef_common.h: improved
* Ctrl/CalcExpr_Parse.cpp etc.: cond. compilation only with #ifndef DEF_NO_StringUSAGE
* Jc/ReflectionJc.c: here the old known warning pointer to int conversion is fixed.
* Test/testAssert.h: also ERROR on exception!

+++ emC/Ctrl/T1_Ctrl_emC.h
@@ -289,7 +289,7 @@ typedef struct Delayf_Ctrl_emC_T {
  */
 INLINE_emC Delayf_Ctrl_emC_s* ctor_Delayf_Ctrl_emC ( ObjectJc* othiz, int size, int sizeMem ) {
 #ifndef __ignoreInCheader_zbnf__
-  ASSERT_emC(sizeMem >= (size-2) * sizeof(float) + sizeof(Delayf_Ctrl_emC_s), "faulty size on Delayf_Ctrl_emC", size, sizeMem); 
+  ASSERT_emC(sizeMem >= (int)((size-2) * sizeof(float) + sizeof(Delayf_Ctrl_emC_s)), "faulty size on Delayf_Ctrl_emC", size, sizeMem); 
   Delayf_Ctrl_emC_s* thiz = C_CAST(Delayf_Ctrl_emC_s*, othiz);
   thiz->nsize = size;
   return thiz;
